# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "week 3  API",
    "description": " API for user contact Information and Temple Information",
    "version": "1.0.0",
    "contact": {
      "name": "Developer"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
    "/user/create": {
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user in the database",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "description": "First name of the user"
                  },
                  "lastName": {
                    "type": "string",
                    "description": "Last name of the user"
                  },
                  "favouriteColor": {
                    "type": "string",
                    "description": "User's favorite color"
                  },
                  "birthday": {
                    "type": "string",
                    "description": "User's birthdate in YYYY-MM-DD format"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "favouriteColor": {
                      "type": "string"
                    },
                    "birthday": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Content can not be empty"
          },
          "500": {
            "description": "Some error occurred while creating the user"
          }
        }
      }
    },
    "/user/getOne/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieve a single user by their ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          },
          {
            "name": "apiKey",
            "in": "header",
            "required": true,
            "description": "API key for authentication",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "lastName": {
                      "type": "string"
                    },
                    "favouriteColor": {
                      "type": "string"
                    },
                    "birthday": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Cannot find user with the specified ID"
          },
          "500": {
            "description": "Error retrieving user"
          }
        }
      }
    },
    "/user/getAll": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieve all users in the database",
        "parameters": [
          {
            "name": "apiKey",
            "in": "header",
            "required": true,
            "description": "API key for authentication",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string"
                      },
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "favouriteColor": {
                        "type": "string"
                      },
                      "birthday": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Some error occurred while retrieving users"
          }
        }
      }
    },
    "/user/update/{id}": {
      "post": {
        "summary": "Update a user",
        "description": "Update a user's details by their ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "favouriteColor": {
                    "type": "string"
                  },
                  "birthday": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Data to update cannot be empty"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Error updating user"
          }
        }
      }
    },
    "/user/delete/{id}": {
      "delete": {
        "summary": "Delete a user",
        "description": "Delete a user by their ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Could not delete user"
          }
        }
      }
    },
    "/user/deleteAll": {
      "delete": {
        "summary": "Delete all users",
        "description": "Delete all users in the database",
        "responses": {
          "200": {
            "description": "All users deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message with the number of deleted users"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error occurred while removing all users"
          }
        }
      }
    },
...
  },
...
}